+++ packages/waku/src/lib/handlers/handler-prd.ts
@@ -2,9 +2,9 @@ import type { EntriesPrd } from '../../server.js';
 import { resolveConfig } from '../config.js';
 import type { Config } from '../config.js';
 import { endStream } from '../utils/stream.js';
-import { renderHtml } from './html-renderer.js';
-import { decodeInput, hasStatusCode, deepFreeze } from './utils.js';
-import { renderRsc } from '../rsc/rsc-renderer.js';
+import { renderHtml } from '../renderers/html-renderer.js';
+import { decodeInput, hasStatusCode, deepFreeze } from '../renderers/utils.js';
+import { renderRsc } from '../renderers/rsc-renderer.js';
 import type { BaseReq, BaseRes, Handler } from './types.js';
 
 export function createHandler<
