+++ packages/waku/src/lib/middleware/connect-dev.ts
@@ -1,5 +1,5 @@
 import { connectWrapper } from './connect-utils.js';
-import { createHandler } from '../rsc/handler-dev.js';
+import { createHandler } from '../handlers/handler-dev.js';
 
 export function connectMiddleware(...args: Parameters<typeof createHandler>) {
   return connectWrapper(createHandler(...args));
