+++ packages/waku/src/lib/middleware/connect-prd.ts
@@ -1,5 +1,5 @@
 import { connectWrapper } from './connect-utils.js';
-import { createHandler } from '../rsc/handler-prd.js';
+import { createHandler } from '../handlers/handler-prd.js';
 
 export function connectMiddleware(...args: Parameters<typeof createHandler>) {
   return connectWrapper(createHandler(...args));
