+++ packages/waku/src/lib/middleware/hono-utils.ts
@@ -1,6 +1,6 @@
 import type { MiddlewareHandler, Context, Env, Input } from 'hono';
 
-import type { BaseReq, BaseRes, Handler } from '../rsc/types.js';
+import type { BaseReq, BaseRes, Handler } from '../handlers/types.js';
 
 const createEmptyReadableStream = () =>
   new ReadableStream({
