+++ packages/waku/src/lib/rsc/worker-impl.ts
@@ -8,9 +8,9 @@ import { createServer as createViteServer } from 'vite';
 import type { EntriesDev } from '../../server.js';
 import type { ResolvedConfig } from '../config.js';
 import { joinPath, fileURLToFilePath } from '../utils/path.js';
-import { hasStatusCode, deepFreeze } from './utils.js';
+import { hasStatusCode, deepFreeze } from '../renderers/utils.js';
 import type { MessageReq, MessageRes, RenderRequest } from './worker-api.js';
-import { renderRsc } from './rsc-renderer.js';
+import { renderRsc } from '../renderers/rsc-renderer.js';
 import { nonjsResolvePlugin } from '../plugins/vite-plugin-nonjs-resolve.js';
 import { rscTransformPlugin } from '../plugins/vite-plugin-rsc-transform.js';
 import { rscReloadPlugin } from '../plugins/vite-plugin-rsc-reload.js';
