+++ packages/waku/src/main.ts
@@ -1,3 +1,4 @@
 export { honoMiddleware } from './lib/middleware/hono-prd.js';
 export { connectMiddleware } from './lib/middleware/connect-prd.js';
-export { createHandler as unstable_createHandler } from './lib/rsc/handler-prd.js';
+
+export { createHandler as unstable_createHandler } from './lib/handlers/handler-prd.js';
